+++ src/components/navbar/LectureNavbar.jsx
@@ -6,14 +6,64 @@ import { Outlet } from "react-router-dom";
 import {
   StyledNav,
   StyledDiv,
-} from "./CommunityRouterStyle.";
-const AdminCommunity = () => {
+} from "./LectureNavbar.Style";
+const AdminLecture = () => {
   const location = useLocation();
   return (
     <div>
       <div>
         <StyledNav>
-          <Link to="/community/notice">
+          <Link to="/lecture/introduce">
+            <StyledDiv
+              active={
+                location.pathname.includes(
+                  "introduce"
+                )
+                  ? "true"
+                  : undefined
+              }
+            >
+              강의소개
+            </StyledDiv>
+          </Link>
+          <Link to="/lecture/order">
+            <StyledDiv
+              active={
+                location.pathname.includes(
+                  "order"
+                )
+                  ? "true"
+                  : undefined
+              }
+            >
+              목차
+            </StyledDiv>
+          </Link>
+          <Link to="/lecture/qna">
+            <StyledDiv
+              active={
+                location.pathname.includes("qna")
+                  ? "true"
+                  : undefined
+              }
+            >
+              QnA
+            </StyledDiv>
+          </Link>
+          <Link to="/lecture/course-review">
+            <StyledDiv
+              active={
+                location.pathname.includes(
+                  "course-review"
+                )
+                  ? "true"
+                  : undefined
+              }
+            >
+              수강평
+            </StyledDiv>
+          </Link>
+          <Link to="/lecture/notice">
             <StyledDiv
               active={
                 location.pathname.includes(
@@ -26,34 +76,10 @@ const AdminCommunity = () => {
               공지사항
             </StyledDiv>
           </Link>
-          <Link to="/community/qna">
-            <StyledDiv
-              active={
-                location.pathname.includes("qna")
-                  ? "true"
-                  : undefined
-              }
-            >
-              QnA
-            </StyledDiv>
-          </Link>
-          <Link to="/community/concern">
-            <StyledDiv
-              active={
-                location.pathname.includes(
-                  "concern"
-                )
-                  ? "true"
-                  : undefined
-              }
-            >
-              고민
-            </StyledDiv>
-          </Link>
         </StyledNav>
       </div>
       <Outlet />
     </div>
   );
 };
-export default AdminCommunity;
+export default AdminLecture;
